Code shuffle to move QgsProxyFeatureSink to its own file
