+++ src/core/qgsfeaturesink.h
@@ -102,42 +102,6 @@ class CORE_EXPORT QgsFeatureSink
 
 Q_DECLARE_OPERATORS_FOR_FLAGS( QgsFeatureSink::Flags )
 
-
-/**
- * \class QgsProxyFeatureSink
- * \ingroup core
- * A simple feature sink which proxies feature addition on to another feature sink.
- *
- * This class is designed to allow factory methods which always return new QgsFeatureSink
- * objects. Since it is not always possible to create an entirely new QgsFeatureSink
- * (e.g. if the feature sink is a layer's data provider), a new QgsProxyFeatureSink
- * can instead be returned which forwards features on to the destination sink. The
- * proxy sink can be safely deleted without affecting the destination sink.
- *
- * \since QGIS 3.0
- */
-class CORE_EXPORT QgsProxyFeatureSink : public QgsFeatureSink
-{
-  public:
-
-    /**
-     * Constructs a new QgsProxyFeatureSink which forwards features onto a destination \a sink.
-     */
-    QgsProxyFeatureSink( QgsFeatureSink *sink );
-    bool addFeature( QgsFeature &feature, QgsFeatureSink::Flags flags = nullptr ) override { return mSink->addFeature( feature, flags ); }
-    bool addFeatures( QgsFeatureList &features, QgsFeatureSink::Flags flags = nullptr ) override { return mSink->addFeatures( features, flags ); }
-    bool addFeatures( QgsFeatureIterator &iterator, QgsFeatureSink::Flags flags = nullptr ) override { return mSink->addFeatures( iterator, flags ); }
-
-    /**
-     * Returns the destination QgsFeatureSink which the proxy will forward features to.
-     */
-    QgsFeatureSink *destinationSink() { return mSink; }
-
-  private:
-
-    QgsFeatureSink *mSink = nullptr;
-};
-
 Q_DECLARE_METATYPE( QgsFeatureSink * )
 
 #endif // QGSFEATURESINK_H
