+++ src/core/qgsproxyfeaturesink.cpp
@@ -1,8 +1,8 @@
 /***************************************************************************
-                         qgsfeaturesink.cpp
-                         ------------------
-    begin                : April 2017
-    copyright            : (C) 2017 by Nyall Dawson
+                         qgsproxyfeaturesink.cpp
+                         ----------------------
+    begin                : April 2020
+    copyright            : (C) 2020 by Nyall Dawson
     email                : nyall dot dawson at gmail dot com
  ***************************************************************************/
 
@@ -15,30 +15,9 @@
  *                                                                         *
  ***************************************************************************/
 
-#include "qgsfeaturestore.h"
+#include "qgsproxyfeaturesink.h"
 
-bool QgsFeatureSink::addFeature( QgsFeature &feature, QgsFeatureSink::Flags flags )
-{
-  QgsFeatureList features;
-  features << feature;
-  bool result = addFeatures( features, flags );
 
-  if ( !( flags & FastInsert ) )
-  {
-    // need to update the passed feature reference to the updated copy from the features list
-    feature = features.at( 0 );
-  }
-  return result;
-}
-
-bool QgsFeatureSink::addFeatures( QgsFeatureIterator &iterator, QgsFeatureSink::Flags flags )
-{
-  QgsFeature f;
-  bool result = true;
-  while ( iterator.nextFeature( f ) )
-  {
-    result = result && addFeature( f, flags );
-  }
-  return result;
-}
-
+QgsProxyFeatureSink::QgsProxyFeatureSink( QgsFeatureSink *sink )
+  : mSink( sink )
+{}
